Hackthon/requirements.txt
streamlit
numpy
tensorflow-cpu
pandas
math
plotly
datetime
scikit-learn